+++ packages/npm/jns42-core/src/foreign/box.ts
@@ -1,6 +1,6 @@
 import { mainFfi } from "../main-ffi.js";
 import { NULL_POINTER, Pointer } from "../utils/index.js";
-import { ForeignObject } from "./wrapper.js";
+import { ForeignObject } from "./foreign-object.js";
 
 export class Box extends ForeignObject {
   public static allocate(value: Pointer | undefined) {
