+++ packages/npm/jns42-core/src/foreign/index.ts
@@ -1,5 +1,6 @@
 export * from "./box.js";
 export * from "./c-string.js";
+export * from "./foreign-object.js";
 export * from "./names-builder.js";
 export * from "./names.js";
 export * from "./naming.js";
@@ -9,4 +10,3 @@ export * from "./schema-item.js";
 export * from "./schema-transform.js";
 export * from "./sized-string.js";
 export * from "./vec-usize.js";
-export * from "./wrapper.js";
