+++ packages/npm/jns42-core/src/foreign/schema-item.ts
@@ -1,5 +1,5 @@
 import { mainFfi } from "../main-ffi.js";
-import { ForeignObject } from "./wrapper.js";
+import { ForeignObject } from "./foreign-object.js";
 
 export class SchemaItem extends ForeignObject {
   protected drop() {
